requirements.txt
alembic==1.8.1
celery==5.0.0
click==8.1.3
elasticsearch==7.13.0
fastapi==0.79.0
gunicorn==20.1.0
httpx==0.23.1
# pandas==1.5.2
pydantic==1.9.2
PyJWT==2.4.0
PyMySQL==1.0.2
python-decouple==3.6
pythondi==1.2.4
redis==4.3.5
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
uvicorn==0.18.2
python-jose==3.3.0
python-multipart==0.0.5
passlib==1.7.4
beautifulsoup4==4.11.2
soupsieve==2.3.2.post1
ldap3
openpyxl
pytest
pytest-asyncio
pytest-mock
pika